// src/common/type.ts
//product
export interface ProductType {
    _id: ID;
    name: string;
    description: string;
    images: Image[];
    isActive: boolean;
    sizes: Size[];
    categoryId: ID;
    createdAt: AtedAt;
    updatedAt: AtedAt;
}

export interface ID {
    $oid: string;
}

export interface AtedAt {
    $date: Date;
}

export interface Image {
    uid: string;
    lastModified: number;
    lastModifiedDate: Date;
    name: string;
    size: number;
    type: string;
    percent: number;
    originFileObj: OriginFileObj;
    status: string;
    response: Response;
}

export interface OriginFileObj {
    uid: string;
}

export interface Response {
    success: boolean;
    urls: string[];
}

export interface Size {
    size: string;
    quantity: number;
    price: number;
    _id: ID;
}

//user
export interface UserType {
    _id: ID | string;
    username: string;
    avatar: string;
    email: string;
    password: string;
    phoneNumber: number;
    address: string;
    roleId: string;
}

export interface ID {
    $oid: string;
}


//Category
export interface CategoryType {
    _id: ID;
    name: string;
    slug: string;
    products: ID[];
}

export interface ID {
    $oid: string;
}


//order
export interface OrderType {
    _id: ID;
    orderCode: string;
    userId: ID;
    fullName: string;
    email: string;
    phoneNumber: string;
    address: string;
    products: Product[];
    transportFee: number;
    totalPrice: number;
    createdAt: AtedAt;
    updatedAt: AtedAt;
}

export interface ID {
    $oid: string;
}

export interface AtedAt {
    $date: Date;
}

export interface Product {
    productId: ID;
    selectedSize: string;
    selectedQuantity: number;
    message: string;
    status: string;
    paymentMethod: string;
    reason: string;
    data: any[];
    _id: ID;
}

//blog
// Generated by https://quicktype.io

export interface BlogType {
    _id: ID;
    title: string;
    content: string;
    author: string;
    imageTitle: string;
    createdAt: AtedAt;
    updatedAt: AtedAt;
}

export interface ID {
    $oid: string;
}

export interface AtedAt {
    $date: Date;
}

//contact 
// Generated by https://quicktype.io

export interface ContactType {
    _id: ID;
    fullName: string;
    email: string;
    phoneNumber: number;
    content: string;
    createdAt: AtedAt;
    updatedAt: AtedAt;
    IsProcessed: string;
}

export interface AtedAt {
    $date: Date;
}